Move original /timeouts command into subdirectory
Forgot to do that apparently.
